Skip secondary repository if not defined
This commit adds checks to ensure secondary package repository is defined
before attempting to add it.

Otherwise, a default execution will end with a following error:
The task includes an option with an undefined variable. The error was:
'koha_secondary_package_repository' is undefined

The error appears to be in
'/home/ansible/KKKAnsible/roles/ansible-role-koha/tasks/main.yml': line 45,
column 3